Factor out cylindrical.
Again, just adding imports on top, and an import from cylindrical
at the bottom of __init__.py.
